+++ StudentsMng/Program.cs
@@ -1,17 +1,19 @@
 using Microsoft.EntityFrameworkCore;
-using StudentManager.Models;
+using StudentsManager.Models;
 using StudentsManager.DAL.Context;
 
 var builder = WebApplication.CreateBuilder(args);
+var db_connection_str = builder.Configuration.GetConnectionString("default");
 
 // Add services to the container.
 builder.Services.AddControllersWithViews();
+builder.Services.AddDbContext<StudentsDB>(
+    options => {
+        options.UseSqlServer(db_connection_str);
+    });
 
-var db_connection_str = builder.Configuration.GetConnectionString("default");
 
-builder.Services.AddDbContext<StudentsDB>(options => options.UseSqlServer(db_connection_str));
 var app = builder.Build();
-
 using (var scope = app.Services.CreateScope())
 {
     var services = scope.ServiceProvider;
