+++ StudentsMng/ViewModels/ErrorViewModel.cs
@@ -1,0 +1,8 @@
+namespace StudentsManager.ViewModels;
+
+public class ErrorViewModel
+{
+    public string? RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+}
