+++ StudentsMng/Views/Shared/Error.cshtml
@@ -1,4 +1,5 @@
-﻿@model ErrorViewModel
+﻿@using StudentsManager.ViewModels;
+@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
