+++ StudentsMng/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using StudentsManager
-@using StudentsManager.Models
+@using StudentsManager.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
